pyproject.toml
[project]
name = "wcpan-drive-feed"
version = "1.0.0"
description = "HTTP server exposing local filesystem as a wcpan.drive feed API"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Wei-Cheng Pan", email = "legnaleurc@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "aiohttp>=3.13.0,<4.0.0",
    "asyncinotify>=4.4.0,<5.0.0",
    "python-magic>=0.4.27,<0.5.0",
    "pymediainfo>=7.0.1,<8.0.0",
    "PyYAML>=6.0.1,<7.0.0",
    "wcpan-logging>=2.1.0,<3.0.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: POSIX :: Linux",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]

[project.scripts]
"wcpan.drive.feed" = "wcpan.drive.feed._main:main"

[project.urls]
source = "https://github.com/legnaleurc/wcpan.drive.feed"

[dependency-groups]
dev = [
    "ruff>=0.15.0,<1.0.0",
    "aiohttp[speedups]>=3.13.0,<4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wcpan"]
